+++ tests/050_conf/001_model/001_init.py
@@ -12,8 +12,8 @@ def test_001_default(settings, sample_project_settings):
 
     assert settings_object._settings == DEFAULT_SETTINGS
 
-    assert settings_object.TARGET_PATH == DEFAULT_SETTINGS['TARGET_PATH']
-    assert settings_object.OUTPUT_STYLES == DEFAULT_SETTINGS['OUTPUT_STYLES']
+    assert settings_object.TARGET_PATH == DEFAULT_SETTINGS["TARGET_PATH"]
+    assert settings_object.OUTPUT_STYLES == DEFAULT_SETTINGS["OUTPUT_STYLES"]
 
 
 def test_002_minimal(settings, sample_project_settings):
@@ -21,8 +21,8 @@ def test_002_minimal(settings, sample_project_settings):
     Very minimalistic settings
     """
     minimal_conf = {
-        'SOURCES_PATH': '.',
-        'TARGET_PATH': './css',
+        "SOURCES_PATH": ".",
+        "TARGET_PATH": "./css",
     }
 
     settings_object = Settings(initial=minimal_conf)
@@ -32,5 +32,5 @@ def test_002_minimal(settings, sample_project_settings):
 
     assert settings_object._settings == attempted
 
-    assert settings_object.TARGET_PATH == attempted['TARGET_PATH']
-    assert settings_object.OUTPUT_STYLES == attempted['OUTPUT_STYLES']
+    assert settings_object.TARGET_PATH == attempted["TARGET_PATH"]
+    assert settings_object.OUTPUT_STYLES == attempted["OUTPUT_STYLES"]