+++ tests/050_conf/002_base_backend/005_clean.py
@@ -4,7 +4,7 @@ from boussole.conf.base_backend import SettingsBackendBase
 
 def test_001(settings, custom_project_settings):
     """
-    Ensure cleaning dont drop anything
+    Ensure cleaning don't drop anything
     """
     backend = SettingsBackendBase(basedir=settings.fixtures_path)